+++ src/asp/Sessions/StereoSessionGdal.h
@@ -201,6 +201,30 @@ namespace asp {
     }
   };
 
+  /// StereoSession instance for processing Pleiades data.
+  class StereoSessionPleiades : public StereoSessionGdal {
+
+  public:
+    StereoSessionPleiades(){}
+    virtual ~StereoSessionPleiades(){}
+    
+    virtual std::string name() const { return "pleiades"; }
+    
+    /// Simple factory function
+    static StereoSession* construct() { return new StereoSessionPleiades; }
+    
+  protected:
+    /// Function to load a camera model of the particular type.
+    virtual boost::shared_ptr<vw::camera::CameraModel>
+    load_camera_model(std::string const& image_file,
+                      std::string const& camera_file,
+                      vw::Vector2 pixel_offset) const{
+      
+      return load_adjusted_model(m_camera_loader.load_pleiades_camera_model(camera_file),
+                                 image_file, camera_file, pixel_offset);
+    }
+  };
+  
 } // End namespace asp
 
 #endif//__STEREO_SESSION_GDAL_H__